refactor: xxx.style.ts 로 설정한 스타일 파일 이름 xxx.styles.ts로 변경
