+++ src/components/mypage/Content.styles.ts
