+++ src/components/mypage/MyFeed.styles.ts
