+++ src/components/pack_feed_writing/HashTag.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
-import { ColumnWrapper } from '@/components/pack_feed_writing/Layout.style';
-import { Desc, TitleStyle } from '@/components/pack_feed_writing/Title.style';
+import { ColumnWrapper } from '@/components/pack_feed_writing/Layout.styles';
+import { Desc, TitleStyle } from '@/components/pack_feed_writing/Title.styles';
 import {
   HashTagBox,
   TagContainer,
@@ -8,7 +8,7 @@ import {
   RemoveBtn,
   InputTag,
   WarningText,
-} from '@/components/pack_feed_writing/HashTag.style';
+} from '@/components/pack_feed_writing/HashTag.styles';
 
 import { validateTag } from '@/utils/validateHashTag';
