+++ src/components/signup/SignupStep.styles.ts
